cope with libcurl being in /usr/lib or /usr/lib64
forgot a space

fail if expected tars aren't built

debug

+++ server_build/build_pkg.sh
@@ -388,9 +388,16 @@ function prep_server {
         fi
 
         # Curl
-        if [ ! -z $CURL_HOME -a -e $CURL_HOME/lib ] ; then
+        if [ ! -z $CURL_HOME -a -e $CURL_HOME/lib -a -e $CURL_HOME/lib/libcurl.so.4 ] ; then
           cp --no-preserve=ownership $CURL_HOME/lib/libcurl.so.4 $IMASERVER_BASE_DIR/${IMA_SVR_INSTALL_PATH}/lib64/.
           cp --no-preserve=ownership $CURL_HOME/lib/libcurl.so.4 $IMASERVER_BASE_DIR/${IMA_SVR_INSTALL_PATH}/debug/lib/.
+        elif [ ! -z $CURL_HOME -a -e $CURL_HOME/lib -a -e $CURL_HOME/lib64/libcurl.so.4 ] ; then
+          cp --no-preserve=ownership $CURL_HOME/lib64/libcurl.so.4 $IMASERVER_BASE_DIR/${IMA_SVR_INSTALL_PATH}/lib64/.
+          cp --no-preserve=ownership $CURL_HOME/lib64/libcurl.so.4 $IMASERVER_BASE_DIR/${IMA_SVR_INSTALL_PATH}/debug/lib/.
+        elif [ ! -z $CURL_HOME -a -e $CURL_HOME/lib ] ; then
+            # oh dead
+            echo "Cant find libcurl either $CURL_HOME/lib/libcurl.so.4 or $CURL_HOME/lib64/libcurl.so.4"
+            exit 12
         fi
 
         # NetSNMP
@@ -855,8 +862,14 @@ function bld_imaproxy_rpm {
             cp --no-preserve=ownership $SSL_HOME/lib/libcrypto.so.1.1 ${PROXY_BASE_DIR}/${IMA_PROXY_INSTALL_PATH}/lib64/.
         fi
 
-        if [ ! -z $CURL_HOME -a -e $CURL_HOME/lib ] ; then
+        if [ ! -z $CURL_HOME -a -e $CURL_HOME/lib -a -e $CURL_HOME/lib/libcurl.so.4 ] ; then
             cp --no-preserve=ownership $CURL_HOME/lib/libcurl.so.4 ${PROXY_BASE_DIR}/${IMA_PROXY_INSTALL_PATH}/lib64/.
+        elif [ ! -z $CURL_HOME -a -e $CURL_HOME/lib -a -e $CURL_HOME/lib64/libcurl.so.4 ] ; then
+            cp --no-preserve=ownership $CURL_HOME/lib64/libcurl.so.4 ${PROXY_BASE_DIR}/${IMA_PROXY_INSTALL_PATH}/lib64/.
+        elif [ ! -z $CURL_HOME -a -e $CURL_HOME/lib ] ; then
+            # oh dead
+            echo "Cant find libcurl either $CURL_HOME/lib/libcurl.so.4 or $CURL_HOME/lib64/libcurl.so.4"
+            exit 11
         fi
     fi
 
@@ -866,8 +879,14 @@ function bld_imaproxy_rpm {
           cp --no-preserve=ownership $MONGOC_HOME/lib64/libmongoc-1.0.so.0 ${PROXY_BASE_DIR}/${IMA_PROXY_INSTALL_PATH}/lib64/.
         fi
 
-        if [ ! -z $MONGOC_HOME -a -e $MONGOC_HOME/lib ] ; then
+        if [ ! -z $MONGOC_HOME -a -e $MONGOC_HOME/lib -a -e $MONGOC_HOME/lib/libcurl.so.4 ] ; then
             cp --no-preserve=ownership $MONGOC_HOME/lib/libcurl.so.4 ${PROXY_BASE_DIR}/${IMA_PROXY_INSTALL_PATH}/lib64/.
+        elif [ ! -z $MONGOC_HOME -a -e $MONGOC_HOME/lib -a -e $MONGOC_HOME/lib64/libcurl.so.4 ] ; then
+            cp --no-preserve=ownership $MONGOC_HOME/lib64/libcurl.so.4 ${PROXY_BASE_DIR}/${IMA_PROXY_INSTALL_PATH}/lib64/.
+        elif [ ! -z $MONGOC_HOME -a -e $MONGOC_HOME/lib ] ; then
+            # oh dead
+            echo "Cant find libcurl either $MONGOC_HOME/lib/libcurl.so.4 or $MONGOC_HOME/lib64/libcurl.so.4"
+            exit 11
         fi
     fi
